ml-challenge-kinase-main/requirements.txt
alabaster==0.7.12
attrs==22.2.0
autopep8==2.0.1
Babel==2.11.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.2.0
dgl-cu116==0.9.1.post1
dglgo==0.0.2
dgllife==0.3.0
docutils==0.19
exceptiongroup==1.0.4
future==0.18.2
hyperopt==0.2.7
idna==3.4
imagesize==1.4.1
importlib-metadata==5.2.0
iniconfig==1.1.1
isort==5.11.3
Jinja2==3.1.2
joblib==1.2.0
littleutils==0.2.2
MarkupSafe==2.1.1
networkx==2.8.8
numpy==1.24.0
numpydoc==1.5.0
ogb==1.3.5
outdated==0.2.2
packaging==22.0
pandas==1.5.2
Pillow==9.3.0
pluggy==1.0.0
psutil==5.9.4
py4j==0.10.9.7
pycodestyle==2.10.0
pydantic==1.10.2
Pygments==2.13.0
pyparsing==3.0.9
pytest==7.2.0
python-dateutil==2.8.2
pytz==2022.7
PyYAML==6.0
rdkit-pypi==2022.9.3
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
scikit-learn==0.24.2
scipy==1.9.3
six==1.16.0
snowballstemmer==2.2.0
Sphinx==5.3.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
threadpoolctl==3.1.0
tomli==2.0.1
torch==1.13.0+cu116
torch-cluster==1.6.0+pt113cu116
torch-geometric==2.2.0
torch-scatter==2.1.0+pt113cu116
torch-sparse==0.6.15+pt113cu116
torch-spline-conv==1.2.1+pt113cu116
torchaudio==0.13.0+cu116
torchvision==0.14.0+cu116
tqdm==4.64.1
typer==0.7.0
typing_extensions==4.4.0
urllib3==1.26.13
zipp==3.11.0
